requirements.txt
git+https://git@github.com/nikett/callgpt.git
numpy
pandas
scipy
Jinja2
faiss-cpu